Bug 1673911 - Overwrite cycle interval in strategies (if legal)
* DRY days validation & split test coverage
* provide test coverage & refactor strategies
